+++ tests/menu/shortcuts/test_exports.py
@@ -2,8 +2,9 @@ import os
 
 import pytest
 
-from REvoDesign.shortcuts.tools.exports import shortcut_dump_fasta_from_struct
-from tests.conftest import TestWorker
+from REvoDesign.shortcuts.tools.exports import (
+    shortcut_dump_fasta_from_struct, shortcut_dump_sidechains)
+from tests.conftest import KeyData, TestWorker
 
 
 @pytest.mark.serial
@@ -40,3 +41,26 @@ def test_shortcut_dump_fasta_from_struct(
     fasta_file_contents = open(expected_fasta_file).readlines()
     assert len([l for l in fasta_file_contents if l.startswith(
         '>')]) == expected_seq_num, f'Expected {expected_seq_num} sequences in {expected_fasta_file}, but got: \n {fasta_file_contents}'
+
+
+@pytest.mark.serial
+def test_shortcut_dump_sidechains(test_worker: TestWorker, KeyDataDuringTests: KeyData):
+    test_worker.load_session_and_check(customized_session=KeyDataDuringTests.evaluate_pse_path)
+
+    sele = 'mt_Q239_3.0'
+    save_dir = f'png/sidechain_dump_{sele}/'
+    shortcut_dump_sidechains(
+        sele=[sele],
+        enabled_only=False,
+        save_dir=save_dir,
+        height=1280,
+        width=1280,
+        dpi=150,
+        ray=True,
+        hide_mesh=True,
+        neighborhood=3,
+        reorient=True,
+        recenter=False,
+    )
+    pngs = [f for f in os.listdir(save_dir) if f.endswith('.png')]
+    assert len(pngs) == 3, f'Expected 3 pngs in {save_dir}, but got: {pngs}'